services/trading-service/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
python-dotenv>=1.0
pydantic>=2.0
pandas>=2.2
numpy>=1.26
sqlalchemy>=2.0
psycopg2-binary>=2.9
requests>=2.31
httpx>=0.27
smartapi-python>=1.3
pyotp>=2.9
logzero>=1.7
websocket-client>=1.6
prometheus-client>=0.20
